samples: initial commit
Initial commit of the first sample provided for the library: ease. It tests easing functions in the graphical environment. Graphical backend for the sample is: raylib. What's more, easing funtion that utilized power function has the exponent element converted to int, instead of float. Following that, 'pow()' function now uses 'const int' instead of 'const size_t' to account for the potential negative value (notice the easing functions). What I've noticed: 'pow()' function should probably be rewritten to really account for the potential negative exponent.

+++ common.h
@@ -83,12 +83,12 @@ static inline float _max(const float f0, const float f1) {
 
 /* ---------------------------------------------------------------------------------------------------- */
 
-static inline float _pow(const float f0, const size_t s0) {
+static inline float _pow(const float f0, const int s0) {
     if (s0 == 0)      { return (1.0); }
     else if (s0 == 1) { return (f0);  }
 
     float value = f0;
-    for (size_t i = 0; i < s0 - 1; i++) {
+    for (int i = 0; i < s0 - 1; i++) {
         value *= f0;
     }
